refactor: move history page files
