+++ src/app/[locale]/history/page.tsx
@@ -1,4 +1,4 @@
-import HistoryPageContent from './HistoryPageContent';
+import HistoryPageContent from '@/components/historyPage/HistoryPageContent';
 
 export default function HistoryPage() {
   return <HistoryPageContent />;
